+++ src/main/java/com/beyond/easycheck/admin/ui/requestbody/UserStatusUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.beyond.easycheck.user.ui.requestbody;
+package com.beyond.easycheck.admin.ui.requestbody;
 
 import com.beyond.easycheck.user.application.domain.UserStatus;
 import jakarta.validation.constraints.NotNull;
